+++ src/extension.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import { markCode } from './commands/markCode';
 import { addTitle } from './commands/addTitle';
 import { navigate } from './commands/navigate';
-import { linkMark } from './commands/linkMark';
+import { connectMark } from './commands/connectMark';
 import { CodeTrailLinkProvider } from './providers/linkProvider';
 import { showGraph } from './commands/showGraph';
 import { initLogger, log } from './utils/logger';
@@ -15,7 +15,7 @@ export function activate(context: vscode.ExtensionContext) {
 		vscode.commands.registerCommand('codeTrail.markCode', markCode),
 		vscode.commands.registerCommand('codeTrail.addTitle', addTitle),
 		vscode.commands.registerCommand('codeTrail.navigate', navigate),
-		vscode.commands.registerCommand('codeTrail.linkMark', linkMark),
+		vscode.commands.registerCommand('codeTrail.connectMark', connectMark),
 		vscode.commands.registerCommand('codeTrail.showGraph', () =>
 			showGraph(context),
 		),